requirements.txt
Flask
gunicornpylibmc
